requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
httpx
google-auth
google-auth-oauthlib
google-api-python-client
